requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.2
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.5
django-allauth==0.41.0
django-bootstrap-datepicker-plus==3.0.5
django-countries==6.1.2
django-crispy-forms==1.9.0
django-extensions==2.2.9
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-datatables==0.5.2
filelock==3.0.12
gunicorn==20.0.4
heroku==0.1.4
idna==2.9
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
pycodestyle==2.5.0
Pygments==2.6.1
pylint==2.4.4
python-dateutil==1.5
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
virtualenv==20.0.18
wcwidth==0.1.9
whitenoise==5.1.0
wrapt==1.11.2
